Rename module from exportdirectory to directory_api_client
