+++ directory_api_client/company.py
@@ -1,4 +1,4 @@
-from exportdirectory.base import BaseAPIClient
+from directory_api_client.base import BaseAPIClient
 
 
 class CompanyAPIClient(BaseAPIClient):
